requirements.txt
asgiref==3.8.1
attrs==23.2.0
autopep8==2.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cockroachdb==5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
h11==0.14.0
idna==3.7
iniconfig==2.0.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.1
pytest-sugar==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.0
termcolor==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
